Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

This date call works with different implementations of date.

Also use UTC to be independent of timezone.

This patch was done while working on reproducible builds for openSUSE.